+++ src/entity/UserFactory.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public interface UserFactory {
 
-    User create(String username, String password, String name, String gender, Double weight, int age,
-                Double height, ArrayList<String> dietaryRestrictions);
+    User create(String username, String password, String name, String gender, Float weight, int age,
+                Float height, ArrayList<String> dietaryRestrictions);
 
 }
